source/pyproject.toml
[project]
name = "digital-event-manager-bot"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.18.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "ruff>=0.11.0",
    "python-logging-loki>=0.3.1"
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
